requirements.txt
defusedxml==0.4.1
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.4
django-appconf==1.0.1
django-compressor==2.0
django-libsass==0.6
django-pipeline==1.6.8
django-pipeline-browserify==0.2.0
django-richtextfield==1.1
django-toolbelt==0.0.1
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
gunicorn==19.4.5
httplib2==0.9.2
libsass==0.11.0
psycopg2==2.6.1
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
static3==0.6.1
whitenoise==3.0
